UPDATE: use long for ids (#93)

+++ src/main/java/mops/presentation/GroupController.java
@@ -36,7 +36,7 @@ public class GroupController {
     @GetMapping(path = "/{groupId}")
     public String getAllFilesOfDirectory(KeycloakAuthenticationToken token,
                                          Model model,
-                                         @PathVariable("groupId") int groupId) {
+                                         @PathVariable("groupId") long groupId) {
         final Account account = AccountUtil.getAccountFromToken(token);
         final List<FileInfo> files = fileService.getAllFilesOfGroup(account, groupId);
         model.addAttribute("files", files); //NOPMD
@@ -55,7 +55,7 @@ public class GroupController {
     @PostMapping(path = "/{groupId}/search")
     public String searchFilesInGroup(KeycloakAuthenticationToken token,
                                      Model model,
-                                     @PathVariable("groupId") int groupId,
+                                     @PathVariable("groupId") long groupId,
                                      @Param("searchQuery") FileQuery query) {
         final Account account = AccountUtil.getAccountFromToken(token);
         final List<FileInfo> files = fileService.searchFilesInGroup(account, groupId, query);
